refactor: removed console and comments
