+++ src/Pages/AuthPages/SignupPage.js
@@ -1,9 +1,8 @@
 import React from 'react'
 import './AuthPage.css'
-import Navbar from '../../Components/Navbar/Navbar'
 import { Link, useNavigate } from 'react-router-dom'
 import { useState } from 'react'
-import FormInput from './AuthComponents/FormInput'
+import { FormInput, Navbar } from './index.js'
 const axios = require('axios');
 
 const SignupPage = () => {
@@ -20,7 +19,6 @@ const SignupPage = () => {
                 '/api/auth/signup',
                 { ...signupData }
             )
-            console.log(res)
             if (res.status === 200 || res.status === 201) {
                 setUserExistError(false)
                 setSignupError(false)
@@ -28,7 +26,6 @@ const SignupPage = () => {
                 navigate("/")
             }
         } catch (error) {
-            console.log(error)
             setSignupError(true)
             if (error.response.status === 422) {
                 setSignupError(false)
