+++ src/Pages/AuthPages/index.js
@@ -1,0 +1,4 @@
+import FormInput from './AuthComponents/FormInput'
+import Navbar from '../../Components/Navbar/Navbar'
+
+export { FormInput, Navbar }
