+++ src/Pages/HomePage/HomePage.js
@@ -26,4 +26,4 @@ const HomePage = () => {
     )
 }
 
-export { HomePage }
+export default HomePage
