+++ src/Pages/index.js
@@ -1,0 +1,6 @@
+import HomePage from './HomePage/HomePage';
+import LoginPage from './AuthPages/LoginPage';
+import LogoutPage from './AuthPages/LogoutPage';
+import SignupPage from './AuthPages/SignupPage';
+
+export { HomePage, LoginPage, LogoutPage, SignupPage }
